+++ actions/rotate.c
@@ -7,8 +7,8 @@ struct image rotate_90(struct image src) {
 
     struct image result_image = image_create(result_width, result_height);
 
-    for (size_t i = 0; i < src.height; i++){
-        for (size_t j = 0; j < src.width; j++){
+    for (uint32_t i = 0; i < src.height; i++){
+        for (uint32_t j = 0; j < src.width; j++){
             result_image.data[j * result_width + result_width-i-1] = src.data[i*src.width + j];
         }
     }
